# [hpc infer] Backout D31360987 and D31293307