Summary:
We see the perf regression on the trunk; After backout these 2 Diffs, the perf recovers.

Back out "Do const_prop before acc tracing":  D31360987

Original commit changeset: 2e5bdc6159d8

Back out "[fx][const_fold] Refactor to use base split module to simplify, and correctly handle non-single-Tensor outputs": D31293307 (https://github.com/pytorch/pytorch/commit/592481a5cc8c6d97c6e5def95789d1d9f6e818d9)

Original commit changeset: 6e283a8c7222

- Current Trunk:
```
375.784834MFLOP per batch item
OverArch, PyTorch, FP16, BS: 2048, TFLOP/s: 53.14, Time per iter: 14.48ms, QPS: 141411.57
OverArch, TensorRT, FP16, BS: 2048, TFLOP/s: 59.62, Time per iter: 12.91ms, QPS: 158656.17
```

Profiling trace: https://fburl.com/perfdoctor/ciw4nktx

- After the backout
```
375.784834MFLOP per batch item
OverArch, PyTorch, FP16, BS: 2048, TFLOP/s: 53.43, Time per iter: 14.40ms, QPS: 142179.48
OverArch, TensorRT, FP16, BS: 2048, TFLOP/s: 97.55, Time per iter: 7.89ms, QPS: 259585.06
```

Profiling trace: https://fburl.com/perfdoctor/eqdxcesv

Test Plan: buck run mode/opt -c python.package_style=inplace -c fbcode.nvcc_arch=a100 //hpc/new/models/ads/benchmarks:ads_dense_benchmark -- over-arch --model-version=23x_3tb --batch-size=2048 2>&1 | tee out.log

Differential Revision: D31497913